+++ src/main/java/com/chong/girl/dao/GirlRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GirlRepository extends JpaRepository<Girl,Integer> {
+public interface GirlRepository extends JpaRepository<Girl,Long> {
 
      List<Girl> findGirlByAge(Integer age);
 }
